% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{bill_action_history}
\alias{bill_action_history}
\title{Get action history for a specific bill}
\usage{
bill_action_history(bill_id)
}
\arguments{
\item{bill_id}{The unique identifier of the bill}
}
\value{
A tibble containing action history
}
\description{
This function retrieves the action history for a specific bill.
}
\examples{
bill_action_history(12345)

}
