% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{bill_details}
\alias{bill_details}
\title{Get detailed information about a specific bill}
\usage{
bill_details(bill_id)
}
\arguments{
\item{bill_id}{The unique identifier of the bill}
}
\value{
A tibble containing detailed bill information
}
\description{
This function retrieves detailed information about a bill by its ID.
}
\examples{
bill_details(12345)

}
