% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{bill_events}
\alias{bill_events}
\title{Get events for a specific bill}
\usage{
bill_events(bill_id)
}
\arguments{
\item{bill_id}{The unique identifier of the bill}
}
\value{
A tibble containing event information
}
\description{
This function retrieves all events for a specific bill.
}
\examples{
bill_events(12345)

}
