% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bill_sheet_bills}
\alias{bill_sheet_bills}
\title{Get bills from a specific bill sheet}
\usage{
bill_sheet_bills(bill_sheet_id)
}
\arguments{
\item{bill_sheet_id}{The unique identifier of the bill sheet}
}
\value{
A tibble of bills contained in the specified bill sheet
}
\description{
This function retrieves all bills contained in a specified bill sheet.
}
\examples{
bill_sheet_bills(12345)

}
