% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bill_sheets}
\alias{bill_sheets}
\title{Get information about bill sheets}
\usage{
bill_sheets()
}
\value{
A tibble of bill sheets containing information like billSheetID, name, summary, counts, etc.
}
\description{
This function retrieves information about all bill sheets for the authenticated user.
}
\examples{
bill_sheets()

}
