% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{bill_sponsors}
\alias{bill_sponsors}
\title{Get sponsors of a specific bill}
\usage{
bill_sponsors(bill_id)
}
\arguments{
\item{bill_id}{The unique identifier of the bill}
}
\value{
A tibble containing sponsor information
}
\description{
This function retrieves all sponsors for a specific bill.
}
\examples{
bill_sponsors(12345)

}
