% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{bill_votes}
\alias{bill_votes}
\title{Get votes for a specific bill}
\usage{
bill_votes(bill_id)
}
\arguments{
\item{bill_id}{The unique identifier of the bill}
}
\value{
A tibble containing voting information
}
\description{
This function retrieves all votes for a specific bill.
}
\examples{
bill_votes(12345)

}
