% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{bills}
\alias{bills}
\title{Get information about bills}
\usage{
bills(search_text = NULL, state_codes = NULL, session_id = NULL)
}
\arguments{
\item{search_text}{The text to search for in bill titles}

\item{state_codes}{The state codes to filter bills by}

\item{session_id}{The session ID to filter bills by}
}
\value{
A list of bills
}
\description{
This function retrieves information about bills based on the specified search text, state codes, and session ID.
}
\examples{
bills(search_text = "education", state_codes = "NY", session_id = 12345)

}
