% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{billtrack_req}
\alias{billtrack_req}
\title{Retrieve data from the BillTrack50 API}
\usage{
billtrack_req(
  path,
  format = "json",
  api_key = Sys.getenv("BILLTRACK50_API_KEY")
)
}
\arguments{
\item{path}{The specific endpoint path to request data from}

\item{format}{The format in which the data should be returned (default is "json")}

\item{api_key}{The API key required to access the BillTrack50 API (default is retrieved from environment variable 'BILLTRACK50_API_KEY')}
}
\value{
A list containing the data retrieved from the API
}
\description{
Retrieve data from the BillTrack50 API
}
\examples{
billtrack_req("sessions/CA")
billtrack_req("bills/?searchText=education&stateCodes=CA&sessionID=12345")
billtrack_req("legislators?legislatorName=John\%20Doe&stateCodes=CA")

}
