% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislators.R
\name{legislator_bills}
\alias{legislator_bills}
\title{Get bills sponsored by a specific legislator}
\usage{
legislator_bills(legislator_id)
}
\arguments{
\item{legislator_id}{The unique identifier of the legislator}
}
\value{
A tibble containing bills sponsored by the legislator
}
\description{
This function retrieves all bills sponsored by a specific legislator.
}
\examples{
legislator_bills(18718)

}
