% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislators.R
\name{legislator_details}
\alias{legislator_details}
\title{Get detailed information about a specific legislator}
\usage{
legislator_details(legislator_id)
}
\arguments{
\item{legislator_id}{The unique identifier of the legislator}
}
\value{
A tibble containing detailed legislator information
}
\description{
This function retrieves detailed information about a legislator by their ID.
}
\examples{
legislator_details(18718)

}
