% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislators.R
\name{legislator_votes}
\alias{legislator_votes}
\title{Get voting history for a specific legislator}
\usage{
legislator_votes(legislator_id)
}
\arguments{
\item{legislator_id}{The unique identifier of the legislator}
}
\value{
A tibble containing the legislator's voting history
}
\description{
This function retrieves all votes cast by a specific legislator.
}
\examples{
legislator_votes(18718)

}
