% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislators.R
\name{legislators}
\alias{legislators}
\title{Get information about legislators}
\usage{
legislators(legislator_name, state_codes)
}
\arguments{
\item{legislator_name}{The name of the legislator to search for}

\item{state_codes}{The state codes to filter legislators by}
}
\value{
A tibble of legislators
}
\description{
This function retrieves information about legislators based on the specified legislator name and state codes.
}
\examples{
legislators("John Doe", "NY")

}
