% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecards.R
\name{scorecard_legislator}
\alias{scorecard_legislator}
\title{Get legislator information for a specific scorecard}
\usage{
scorecard_legislator(scorecard_id, legislator_id)
}
\arguments{
\item{scorecard_id}{The unique identifier of the scorecard}

\item{legislator_id}{The unique identifier of the legislator}
}
\value{
A tibble containing legislator information and category scores
}
\description{
This function retrieves information about a specific legislator in a scorecard.
}
\examples{
scorecard_legislator(12345, 18718)

}
