% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecards.R
\name{scorecard_legislator_votes}
\alias{scorecard_legislator_votes}
\title{Get legislator votes for a specific scorecard}
\usage{
scorecard_legislator_votes(scorecard_id, legislator_id)
}
\arguments{
\item{scorecard_id}{The unique identifier of the scorecard}

\item{legislator_id}{The unique identifier of the legislator}
}
\value{
A tibble containing voting information
}
\description{
This function retrieves voting information for a specific legislator in a scorecard.
}
\examples{
scorecard_legislator_votes(12345, 18718)

}
