% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecards.R
\name{scorecard_legislators}
\alias{scorecard_legislators}
\title{Get all legislators for a specific scorecard}
\usage{
scorecard_legislators(scorecard_id, category_id = NULL)
}
\arguments{
\item{scorecard_id}{The unique identifier of the scorecard}

\item{category_id}{Optional category ID to filter legislation}
}
\value{
A tibble containing legislator information
}
\description{
This function retrieves information about all legislators in a scorecard.
}
\examples{
scorecard_legislators(12345)
scorecard_legislators(12345, category_id = 67)

}
