% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecards.R
\name{scorecards}
\alias{scorecards}
\title{Get information about scorecards}
\usage{
scorecards()
}
\value{
A tibble of scorecards containing information like scorecardID, name, categories, etc.
}
\description{
This function retrieves information about all scorecards for the authenticated user.
}
\examples{
scorecards()

}
