% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sessions}
\alias{sessions}
\title{Get information about sessions}
\usage{
sessions(state = NULL)
}
\arguments{
\item{state}{The state for which to retrieve session information}
}
\value{
A list of sessions
}
\description{
This function retrieves information about sessions based on the specified state.
}
\examples{
sessions("NY")

}
