% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bills.R
\name{state_bill_progress}
\alias{state_bill_progress}
\title{Get bill progress information by state}
\usage{
state_bill_progress(state_codes = NULL)
}
\arguments{
\item{state_codes}{Optional state codes to filter progress information}
}
\value{
A tibble containing bill progress information
}
\description{
This function retrieves information about bill progress in specified states.
}
\examples{
state_bill_progress()
state_bill_progress("CA")

}
